[automation] Rename port-openers to purchase-crackers
Less awkward and more specific.
