+++ src/automation/purchase-crackers.ts
